requirements.txt
pyzotero==1.5.18
PyMuPDF==1.23.26
pdfplumber==0.10.4
openai==1.98.0
google-generativeai>=0.5.0
pydantic==2.5.3
pydantic-core==2.14.6
jinja2==3.1.3
python-dotenv==1.0.1
tqdm==4.66.2